[add] 회원가입 약관 컴포넌트화
